requirements.txt
fastapi>=0.100.0,<0.110.0
uvicorn>=0.23.0,<0.24.0
python-dotenv>=0.19.0
relevanceai>=10.1.0
pydantic>=2.8.2
python-multipart>=0.0.5